Add files via upload
compute_wcs_mse2.pdf contains MSE calculations for humans' responses.
The CSV files are the result of running this code.
I then imported the CSV files on R Studio and performed data cleaning and created 2 figures based on the cleaned data. These figures can also be found on the report under the 'Human Reconstruction' section of the results.